breaking up main into separate packages and class files
